refactor(rust): crate::software::model::{RegistrationParams,RegistrationInfo}
move them from web.rs for use by ProductHTTPClient
