+++ rust/agama-server/src/software/web.rs
@@ -12,11 +12,12 @@ use crate::{
         Event,
     },
 };
+
 use agama_lib::{
     error::ServiceError,
-    product::{proxies::RegistrationProxy, Product, ProductClient, RegistrationRequirement},
+    product::{proxies::RegistrationProxy, Product, ProductClient},
     software::{
-        model::SoftwareConfig,
+        model::{RegistrationInfo, RegistrationParams, SoftwareConfig},
         proxies::{Software1Proxy, SoftwareProductProxy},
         Pattern, SelectedBy, SoftwareClient, UnknownSelectedBy,
     },
@@ -229,19 +230,6 @@ async fn products(State(state): State<SoftwareState<'_>>) -> Result<Json<Vec<Pro
     Ok(Json(products))
 }
 
-/// Information about registration configuration (product, patterns, etc.).
-#[derive(Clone, Serialize, Deserialize, utoipa::ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct RegistrationInfo {
-    /// Registration key. Empty value mean key not used or not registered.
-    key: String,
-    /// Registration email. Empty value mean email not used or not registered.
-    email: String,
-    /// if registration is required, optional or not needed for current product.
-    /// Change only if selected product is changed.
-    requirement: RegistrationRequirement,
-}
-
 /// returns registration info
 ///
 /// * `state`: service state.
@@ -265,15 +253,6 @@ async fn get_registration(
     Ok(Json(result))
 }
 
-/// Software service configuration (product, patterns, etc.).
-#[derive(Clone, Serialize, Deserialize, utoipa::ToSchema)]
-pub struct RegistrationParams {
-    /// Registration key.
-    key: String,
-    /// Registration email.
-    email: String,
-}
-
 #[derive(Clone, Serialize, Deserialize, utoipa::ToSchema)]
 pub struct FailureDetails {
     /// ID of error. See dbus API for possible values
@@ -281,6 +260,7 @@ pub struct FailureDetails {
     /// human readable error string intended to be displayed to user
     message: String,
 }
+
 /// Register product
 ///
 /// * `state`: service state.
